V5.5

+++ FarzamTEWebsite/Models/Transaction_Statistics_M.cs
@@ -14,23 +14,23 @@
         public int? BOBT_saham_Online { get; set; }
         public int? BOBT_saham_Normal { get; set; }
         public int? BOBT_saham_Algorithm { get; set; }
-        public int? BOBT_Total_Value { get; set; }
+        public long? BOBT_Total_Value { get; set; }
         public int? FI_Brokerage_Station { get; set; }
         public int? FI_Online_Normal { get; set; }
         public int? FI_Online_Group { get; set; }
         public int? FI_Online_Other { get; set; }
-        public int? FI_Total_Value { get; set; }
-        public int? BOBT_AND_FI_Total_Value { get; set; }
+        public long? FI_Total_Value { get; set; }
+        public long? BOBT_AND_FI_Total_Value { get; set; }
         public int? BKI_Physical { get; set; }
         public int? BKI_Self { get; set; }
         public int? BKI_Ati { get; set; }
         public int? BKI_Ekhtiar { get; set; }
-        public int? BKI_Total_Value { get; set; }
+        public long? BKI_Total_Value { get; set; }
         public int? BEI_Physical { get; set; }
         public int? BEI_Moshtaghe { get; set; }
         public int? BEI_Other { get; set; }
-        public int? BEI_Total_Value { get; set; }
-        public int? All_Total_Value { get; set; }
+        public long? BEI_Total_Value { get; set; }
+        public long? All_Total_Value { get; set; }
         public string? Date_Monthly { get; set; }
     }
 }
